Move number reading functions into basic.py
